Rename[Module]: 분리 모듈
모듈을 레트로핏 모듈과 데이터 모듈로 분리함
